More refactoring to fix lazy vs eager changes.
